+++ src/components/Dashboard.jsx
@@ -18,10 +18,15 @@ import Link from '@material-ui/core/Link';
 import MenuIcon from '@material-ui/icons/Menu';
 import ChevronLeftIcon from '@material-ui/icons/ChevronLeft';
 import NotificationsIcon from '@material-ui/icons/Notifications';
-import { mainListItems, secondaryListItems } from './listItems';
+import { mainListItems, secondaryListItems, secondaryListItem } from './listItems';
 import Chart from './Chart';
 import Deposits from './Deposits';
 import Orders from './Orders';
+import ListItem from '@material-ui/core/ListItem';
+import ListItemIcon from '@material-ui/core/ListItemIcon';
+import ListItemText from '@material-ui/core/ListItemText';
+import AssignmentIcon from '@material-ui/icons/Assignment';
+
 
 function Copyright() {
   return (
@@ -53,6 +58,7 @@ const useStyles = makeStyles((theme) => ({
     ...theme.mixins.toolbar,
   },
   appBar: {
+    // backgroundColor: "#FFFFFF",
     zIndex: theme.zIndex.drawer + 1,
     transition: theme.transitions.create(['width', 'margin'], {
       easing: theme.transitions.easing.sharp,
@@ -77,6 +83,7 @@ const useStyles = makeStyles((theme) => ({
     flexGrow: 1,
   },
   drawerPaper: {
+    // color: "#E4E4E4",
     position: 'relative',
     whiteSpace: 'nowrap',
     width: drawerWidth,
@@ -117,7 +124,10 @@ const useStyles = makeStyles((theme) => ({
   },
 }));
 
-export default function Dashboard() {
+export default function Dashboard(props) {
+
+  const { user, handleLogOut } = props;
+
   const classes = useStyles();
   const [open, setOpen] = React.useState(true);
   const handleDrawerOpen = () => {
@@ -143,7 +153,13 @@ export default function Dashboard() {
             <MenuIcon />
           </IconButton>
           <Typography component="h1" variant="h6" color="inherit" noWrap className={classes.title}>
-            Dashboard
+            Dashboard {user.email}
+          </Typography>
+          <Typography component="h1" variant="h6" color="inherit" noWrap className>
+            Account
+          </Typography>
+          <Typography component="h1" variant="h6" color="inherit" noWrap className>
+            Logout
           </Typography>
           <IconButton color="inherit">
             <Badge badgeContent={4} color="secondary">
@@ -167,7 +183,16 @@ export default function Dashboard() {
         <Divider />
         <List>{mainListItems}</List>
         <Divider />
-        <List>{secondaryListItems}</List>
+        <List>
+
+        <ListItem button onClick={handleLogOut}>
+        <ListItemIcon>
+          {/* <AssignmentIcon /> */}
+          </ListItemIcon>
+          <ListItemText primary="Sign Out" />
+        </ListItem>
+
+        </List>
       </Drawer>
       <main className={classes.content}>
         <div className={classes.appBarSpacer} />